+++ src/custom_requests.ml
@@ -73,3 +73,54 @@ module Type_enclosing = struct
   let request =
     { meth = ocamllsp_prefixed "typeEnclosing"; encode_params; decode_response }
 end
+
+module Type_search = struct
+  type type_search_result =
+    { name : string
+    ; typ : string
+    ; loc : Range.t
+    ; doc : string option
+    ; cost : int
+    ; constructible : string
+    }
+
+  type params =
+    { uri : Uri.t
+    ; position : Position.t
+    ; limit : int
+    ; query : string
+    ; with_doc : bool
+    }
+
+  type response = type_search_result list
+
+  let encode_params { uri; position; limit; query; with_doc } =
+    let open Jsonoo.Encode in
+    let uri =
+      ("textDocument", object_ [ ("uri", string @@ Uri.toString uri ()) ])
+    in
+    let position = ("position", Position.json_of_t position) in
+    let query = ("query", string query) in
+    let limit = ("limit", int limit) in
+    let with_doc = ("with_doc", bool with_doc) in
+    object_ [ uri; position; query; limit; with_doc ]
+
+  let decode_response response =
+    let open Jsonoo.Decode in
+    let decode_res response =
+      let name = field "name" string response in
+      let typ = field "typ" string response in
+      let loc = field "loc" Range.t_of_json response in
+      let doc = try_optional (field "doc" string) response in
+      let cost = field "cost" int response in
+      let constructible = field "constructible" string response in
+      { name; typ; loc; doc; cost; constructible }
+    in
+    list decode_res response
+
+  let make ~uri ~position ~limit ~query ~with_doc () =
+    { uri; position; limit; query; with_doc }
+
+  let request =
+    { meth = ocamllsp_prefixed "typeSearch"; encode_params; decode_response }
+end