+++ src/treeview_commands.ml
@@ -43,7 +43,34 @@ let terminal_item =
   Vscode.TreeItem.set_command item command;
   item
 
-let items = [ select_sandbox_item; terminal_item ]
+let perform_type_search =
+  let icon =
+    `LightDark
+      Vscode.TreeItem.LightDarkIcon.
+        { light =
+            `String (Path.asset "document-search-light.svg" |> Path.to_string)
+        ; dark =
+            `String (Path.asset "document-search-dark.svg" |> Path.to_string)
+        }
+  in
+  let label =
+    `TreeItemLabel
+      (Vscode.TreeItemLabel.create
+         ~label:"Search a value by type or polarity"
+         ())
+  in
+  let item = Vscode.TreeItem.make_label ~label () in
+  let command =
+    Vscode.Command.create
+      ~title:"Search a value by type or polarity"
+      ~command:"ocaml.search-by-type"
+      ()
+  in
+  Vscode.TreeItem.set_iconPath item icon;
+  Vscode.TreeItem.set_command item command;
+  item
+
+let items = [ select_sandbox_item; terminal_item; perform_type_search ]
 
 let getTreeItem ~element = `Value element